Move discretization format to a separate file and add tests
